requirements.txt
# dependiences
Django
djangorestframework
markdown
django-filter
drf-spectacular
djangorestframework-simplejwt
django-rest-url-filter